Renaming Interfaces, changing hitcount to int

+++ src/main/java/net/dontdrinkandroot/cache/metadata/MetaData.java
@@ -59,7 +59,7 @@ public interface MetaData extends Serializable {
 	/**
 	 * Get the number of hits.
 	 */
-	long getHitCount();
+	int getHitCount();
 
 
 	/**
